refactor(decks): extract page management logic into DeckPagesViewModel
